add fake_reader.py test=develop
