+++ python/paddle/fluid/tests/unittests/test_eager_deletion_dynamic_rnn_base.py
@@ -23,24 +23,7 @@ import paddle.fluid.core as core
 import paddle.fluid as fluid
 from paddle.fluid import compiler
 import numpy as np
-
-
-def fake_reader_creator(word_dict_size,
-                        sample_num,
-                        lower_seq_len=100,
-                        upper_seq_len=200,
-                        class_dim=2):
-    def __reader__():
-        for _ in six.moves.range(sample_num):
-            length = np.random.random_integers(
-                low=lower_seq_len, high=upper_seq_len, size=[1])[0]
-            ids = np.random.random_integers(
-                low=0, high=word_dict_size - 1, size=[length]).astype('int64')
-            label = np.random.random_integers(
-                low=0, high=class_dim - 1, size=[1]).astype('int64')[0]
-            yield ids, label
-
-    return __reader__
+from fake_reader import fake_imdb_reader
 
 
 def train(network, use_cuda, use_parallel_executor, batch_size=32, pass_num=2):
@@ -55,7 +38,7 @@ def train(network, use_cuda, use_parallel_executor, batch_size=32, pass_num=2):
         return
 
     word_dict_size = 5147
-    reader = fake_reader_creator(word_dict_size, batch_size * 40)
+    reader = fake_imdb_reader(word_dict_size, batch_size * 40)
     train_reader = paddle.batch(reader, batch_size=batch_size)
 
     data = fluid.layers.data(
